+++ assets/release_notes.md
@@ -1,8 +1,11 @@
 ### 0.14.0 - November 2023
 ---
 
+- Adds support for Sales Orders
 - Adds option to pause and resume barcode scanning with camera
 - Adds option for "single shot" barcode scanning with camera
+- Fixes bug when removing entire quantity of a stock item
+
 
 ### 0.13.0 - October 2023
 ---