+++ assets/release_notes.md
@@ -5,6 +5,7 @@
 - Adds option to pause and resume barcode scanning with camera
 - Adds option for "single shot" barcode scanning with camera
 - Fixes bug when removing entire quantity of a stock item
+- Add line items to purchase orders directly from the app
 
 
 ### 0.13.0 - October 2023