+++ assets/release_notes.md
@@ -6,6 +6,7 @@
 - Adds option for "single shot" barcode scanning with camera
 - Fixes bug when removing entire quantity of a stock item
 - Add line items to purchase orders directly from the app
+- Add line items to purchase order using barcode scanner
 
 
 ### 0.13.0 - October 2023