+++ assets/release_notes.md
@@ -1,4 +1,4 @@
-### 0.14.0 - November 2023
+### 0.14.0 - December 2023
 ---
 
 - Adds support for Sales Orders